Speed, power is modelled as exponential curve

+++ ISO 19030/createSpeedPower.sql
@@ -17,7 +17,7 @@ BEGIN
 								Trim DOUBLE(10, 5),
 								Displacement DOUBLE(10, 1),
 								Propulsive_Efficiency DOUBLE(10, 5),
-								Speed DOUBLE(10, 5), /* Knots */
-								Power DOUBLE(10, 5), /* kW */
+								Speed DOUBLE(10, 5), /* m/s */
+								Power DOUBLE(12, 5), /* kW */
 								CONSTRAINT UniqueDispTrimSpeed UNIQUE(ModelID, Displacement, Trim, Speed));
 END;
